restructured CHC frontend for backward transformation
