+++ subprojects/frontends/chc-frontend/src/main/java/hu/bme/mit/theta/frontend/chc/ChcBackwardXcfaBuilder.java
@@ -1,0 +1,12 @@
+package hu.bme.mit.theta.frontend.chc;
+
+import hu.bme.mit.theta.chc.frontend.dsl.gen.CHCBaseVisitor;
+import hu.bme.mit.theta.chc.frontend.dsl.gen.CHCParser;
+import hu.bme.mit.theta.xcfa.model.XCFA;
+
+public class ChcBackwardXcfaBuilder extends CHCBaseVisitor<Object> implements ChcXcfaBuilder {
+    @Override
+    public XCFA.Builder buildXcfa(CHCParser parser) {
+        throw new UnsupportedOperationException("Backwards transformation is not implemented yet."); // TODO
+    }
+}
