+++ subprojects/xcfa/xcfa-cli/src/main/java/hu/bme/mit/theta/xcfa/cli/XcfaCli.java
@@ -42,7 +42,8 @@ import hu.bme.mit.theta.common.OsHelper;
 import hu.bme.mit.theta.common.logging.ConsoleLogger;
 import hu.bme.mit.theta.common.logging.Logger;
 import hu.bme.mit.theta.core.decl.VarDecl;
-import hu.bme.mit.theta.frontend.chc.CHCFrontend;
+import hu.bme.mit.theta.frontend.chc.ChcFrontend;
+import hu.bme.mit.theta.frontend.chc.ChcUtils;
 import hu.bme.mit.theta.frontend.transformation.ArchitectureConfig;
 import hu.bme.mit.theta.frontend.transformation.grammar.function.FunctionVisitor;
 import hu.bme.mit.theta.frontend.transformation.model.statements.CProgram;
@@ -203,8 +204,11 @@ public class XcfaCli {
 	@Parameter(names = "--precheck", description = "Perform a pre-check when refining a multithreaded program for possibly higher efficiency", arity = 1)
 	boolean preCheck = true;
 
-	@Parameter(names = "--chc", description = "Parses the input as a Constrained Horn Clause in the CHC-comp format.")
+	@Parameter(names = "--chc", description = "Parse the input as a Constrained Horn Clause in the CHC-comp format")
 	boolean chc = false;
+
+	@Parameter(names="--chc-transformation", description = "Direction of transformation from CHC to XCFA")
+	ChcFrontend.ChcTransformation chcTransformation = ChcFrontend.ChcTransformation.FORWARD;
 
 	@Parameter(names = "--algorithm", description = "Algorithm to use when solving multithreaded programs")
 	XcfaConfigBuilder.Algorithm algorithm = XcfaConfigBuilder.Algorithm.SINGLETHREAD;
@@ -293,7 +297,9 @@ public class XcfaCli {
 			try {
 				final CharStream input = CharStreams.fromStream(new FileInputStream(this.input));
 				if (chc) {
-					CHCFrontend chcFrontend = new CHCFrontend();
+					ChcFrontend.chcTransformation = chcTransformation;
+					ChcUtils.setCharStream(input);
+					ChcFrontend chcFrontend = new ChcFrontend();
 					xcfaBuilder = chcFrontend.buildXcfa(input);
 				} else {
 					final CLexer lexer = new CLexer(input);
